+++ src/components/textarea/textarea.tsx
@@ -1,6 +1,5 @@
 import { Component, Element, Event, EventEmitter, Method, Prop, State, Watch, h } from '@stencil/core';
 import { hasSlot } from '../../utilities/slot';
-import ResizeObserver from 'resize-observer-polyfill';
 
 let id = 0;