+++ docs/getting-started/changelog.md
@@ -15,6 +15,8 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Fixed a bug where `sl-dialog` and `sl-drawer` would emit the `sl-hide` event twice [#275](https://github.com/shoelace-style/shoelace/issues/275)
 - Fixed a bug where calling `event.preventDefault()` on certain form elements wouldn't prevent `sl-form` from submitting [#277](https://github.com/shoelace-style/shoelace/issues/277)
 - Restyled `sl-spinner` so the track is visible and the indicator is smaller.
+- Removed `resize-observer-polyfill` in favor of `@types/resize-observer-browser` since all target browsers support `ResizeObserver`
+
 ## 2.0.0-beta.23
 
 - Added `sl-format-number` component